+++ __tests__/workers/notifications/sourcePostModerationApprovedNotification.ts
@@ -1,5 +1,5 @@
 import { DataSource } from 'typeorm';
-import worker from '../../../src/workers/notifications/sourcePostModerationRejected';
+import worker from '../../../src/workers/notifications/sourcePostModerationApprovedNotification';
 import createOrGetConnection from '../../../src/db';
 import {
   Post,
@@ -38,19 +38,21 @@ describe('SourcePostModerationSubmitted', () => {
     expect(registeredWorker).toBeDefined();
   });
 
-  it('should not send notification when status is not rejected', async () => {
-    const postApproved = {
+  it('should not send notification when status is not approved', async () => {
+    const postRejected = {
+      postId: 'p1',
       sourceId: 'a',
       createdById: '2',
-      status: SourcePostModerationStatus.Approved,
+      status: SourcePostModerationStatus.Rejected,
     };
 
-    const approved = await invokeNotificationWorker(worker, {
-      post: postApproved,
+    const rejected = await invokeNotificationWorker(worker, {
+      post: postRejected,
     });
-    expect(approved).toBeUndefined();
+    expect(rejected).toBeUndefined();
 
     const postPending = {
+      postId: 'p1',
       sourceId: 'a',
       createdById: '2',
       status: SourcePostModerationStatus.Pending,
@@ -64,9 +66,10 @@ describe('SourcePostModerationSubmitted', () => {
 
   it('should send notification to author only', async () => {
     const post = {
+      postId: 'p1',
       sourceId: 'a',
       createdById: '2',
-      status: SourcePostModerationStatus.Rejected,
+      status: SourcePostModerationStatus.Approved,
     };
     await con
       .getRepository(Source)
@@ -82,15 +85,17 @@ describe('SourcePostModerationSubmitted', () => {
     const ctx = result[0].ctx as NotificationPostModerationContext;
 
     expect(result.length).toEqual(1);
-    expect(result[0].type).toEqual('source_post_rejected');
+    expect(result[0].type).toEqual('source_post_approved');
+    expect(ctx.post.id).toEqual('p1');
     expect(ctx.userIds).toEqual(['2']);
   });
 
   it('should not send notification other members', async () => {
     const post = {
+      postId: 'p1',
       sourceId: 'a',
       createdById: '2',
-      status: SourcePostModerationStatus.Rejected,
+      status: SourcePostModerationStatus.Approved,
     };
     await con
       .getRepository(Source)
@@ -120,7 +125,8 @@ describe('SourcePostModerationSubmitted', () => {
     const ctx = result[0].ctx as NotificationPostModerationContext;
 
     expect(result.length).toEqual(1);
-    expect(result[0].type).toEqual('source_post_rejected');
+    expect(result[0].type).toEqual('source_post_approved');
+    expect(ctx.post.id).toEqual(post.postId);
     expect(ctx.userIds).toEqual(['2']);
     const ownerOnly = ctx.userIds.every((id) => id === '2');
     expect(ownerOnly).toBeTruthy();
